requirements.txt
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
colory-pprint==1.5.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
iniconfig==2.3.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
termcolor==3.2.0
uritemplate==4.2.0
urllib3==2.5.0
